lambda-durable-functions/strands-research-agent/src/requirements.txt
aws-durable-execution-sdk-python
strands-agents>=1.0.0
boto3
